requirements.txt
beautifulsoup4==4.11.1
selenium==4.5.0
#python==3.9.0
pyyaml == 6.0
lxml==4.9.2